+++ src/components/editor/text/text-editor.tsx
